Convert SpatialPointsFrame.partition_and_write static method to datashader.spatial.points.to_parquet function
Convert SpatialPointsFrame.from_parquet static method
to datashader.spatial.points.read_parquet function

Also, rename `filename` to `path` to match signature of corresponding
dask.dataframe functions.
